refactor: move order utils
