+++ nicegold_v5/utils.py
@@ -5,6 +5,7 @@ import sys
 import logging
 import subprocess
 import json
+import inspect
 from multiprocessing import cpu_count
 from datetime import datetime
 import time
@@ -277,6 +278,79 @@ def sanitize_price_columns(df: pd.DataFrame) -> pd.DataFrame:
 def validate_indicator_inputs(df, min_rows=100):
     from .entry import validate_indicator_inputs as _f
     return _f(df, min_rows=min_rows)
+
+
+def apply_order_costs(
+    entry: float,
+    sl: float,
+    tp1: float,
+    tp2: float,
+    lot: float,
+    direction: str,
+    spread_value: float = 0.2,
+    slippage_max: float = 0.3,
+    commission_per_lot: float = 0.10,
+) -> tuple[float, float, float, float, float]:
+    """คำนวณต้นทุนคำสั่งและปรับราคาเข้า"""
+    spread_half = spread_value / 2
+    slippage = np.random.uniform(-slippage_max, slippage_max)
+    if direction == "buy":
+        entry_adj = entry + spread_half + slippage
+    else:
+        entry_adj = entry - spread_half - slippage
+    commission = 2 * commission_per_lot * lot * 100
+    return entry_adj, sl, tp1, tp2, commission
+
+
+def ensure_buy_sell(
+    trades_df: pd.DataFrame,
+    df: pd.DataFrame,
+    simulate_fn,
+    fold: int | None = None,
+    outdir: str | None = None,
+) -> pd.DataFrame:
+    """บังคับให้มีทั้ง BUY และ SELL เพื่อป้องกันความลำเอียง"""
+    sides = trades_df.get("side", trades_df.get("type", pd.Series(dtype=str))).str.lower()
+    has_buy = "buy" in sides.values
+    has_sell = "sell" in sides.values
+    if has_buy and has_sell:
+        return trades_df
+
+    logger.info("[ensure_buy_sell] Missing BUY/SELL in fold %s → inject", fold)
+    if "percentile_threshold" in inspect.signature(simulate_fn).parameters:
+        trades_df2 = simulate_fn(df, percentile_threshold=1)
+    else:
+        trades_df2 = simulate_fn(df)
+
+    sides2 = trades_df2.get("side", trades_df2.get("type", pd.Series(dtype=str))).str.lower()
+    has_buy2 = "buy" in sides2.values
+    has_sell2 = "sell" in sides2.values
+
+    if not has_buy2 or not has_sell2:
+        dummy_row = {
+            "entry_time": df.index[0] if len(df.index) else 0,
+            "exit_time": df.index[0] if len(df.index) else 0,
+            "side": "buy",
+            "pnl": 0.0,
+            "fold": 1,
+        }
+        if not has_buy2:
+            trades_df2 = pd.concat([trades_df2, pd.DataFrame([dummy_row])], ignore_index=True)
+        if not has_sell2:
+            dummy_row["side"] = "sell"
+            trades_df2 = pd.concat([trades_df2, pd.DataFrame([dummy_row])], ignore_index=True)
+
+    sides3 = trades_df2.get("side", trades_df2.get("type", pd.Series(dtype=str))).str.lower()
+    has_buy3 = "buy" in sides3.values
+    has_sell3 = "sell" in sides3.values
+    if not (has_buy3 and has_sell3) and outdir:
+        os.makedirs(outdir, exist_ok=True)
+        label = f"fold_{fold}" if fold is not None else "final"
+        out_path = os.path.join(outdir, f"missing_side_{label}.csv")
+        pd.DataFrame().to_csv(out_path, index=False)
+        logger.info("[ensure_buy_sell] Exported zero-trade marker → %s", out_path)
+
+    return trades_df2
 
 def ensure_logs_dir(path: str = "logs") -> None:
     """สร้างไดเรกทอรี logs หากยังไม่มี หรือลบไฟล์ที่ชื่อชนกัน"""
